李锦佳：1.修改了定时器用于定时检查状态 2.增加了辅助类用于根据不同的回款方式进行不同的算法 3.辅助类内增加了bigdecimal的加减乘除 3.增加了回款计划

+++ yrb-common/src/main/java/com/lixiang/ssm/utils/InterestObject.java
@@ -1,6 +1,7 @@
 package com.lixiang.ssm.utils;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -19,10 +20,10 @@ public class InterestObject implements Serializable{
 	private Integer periods;
 	
 	//本金
-	private Double capital;
+	private BigDecimal capital;
 	
 	//利率
-	private Double interest;
+	private BigDecimal interest;
 	
 	//还款时间
 	private Date paymentDate;
@@ -35,19 +36,19 @@ public class InterestObject implements Serializable{
 		this.periods = periods;
 	}
 
-	public Double getCapital() {
+	public BigDecimal getCapital() {
 		return capital;
 	}
 
-	public void setCapital(Double capital) {
+	public void setCapital(BigDecimal capital) {
 		this.capital = capital;
 	}
 
-	public Double getInterest() {
+	public BigDecimal getInterest() {
 		return interest;
 	}
 
-	public void setInterest(Double interest) {
+	public void setInterest(BigDecimal interest) {
 		this.interest = interest;
 	}
